Ajuste Query Visão Detalhe

+++ nds-model/src/main/java/br/com/abril/nds/dto/VisaoEstoqueDTO.java
@@ -2,6 +2,7 @@ package br.com.abril.nds.dto;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import br.com.abril.nds.util.CurrencyUtil;
 import br.com.abril.nds.util.export.Export;
@@ -17,10 +18,10 @@ public class VisaoEstoqueDTO implements Serializable{
 	private String estoque;
 	
 	@Export(label = "Produtos", alignment=Alignment.CENTER, exhibitionOrder = 2)
-	private Long produtos;
+	private BigInteger produtos;
 	
 	@Export(label = "Exemplares", alignment=Alignment.CENTER, exhibitionOrder = 3)
-	private Long exemplares;
+	private BigInteger exemplares;
 	
 	private BigDecimal valor;
 	
@@ -37,19 +38,19 @@ public class VisaoEstoqueDTO implements Serializable{
 		this.estoque = estoque;
 	}
 
-	public Long getProdutos() {
+	public BigInteger getProdutos() {
 		return produtos;
 	}
 
-	public void setProdutos(Long produtos) {
+	public void setProdutos(BigInteger produtos) {
 		this.produtos = produtos;
 	}
 
-	public Long getExemplares() {
+	public BigInteger getExemplares() {
 		return exemplares;
 	}
 
-	public void setExemplares(Long exemplares) {
+	public void setExemplares(BigInteger exemplares) {
 		this.exemplares = exemplares;
 	}
